requirements.txt
flask
pymongo
redis
python-dotenv
Flask-Session
requests
bson